# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Chat App API
  description: API documentation for the Real-time Chat Application Backend
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
tags:
  - name: Authentication
    description: Endpoints for user registration and login
  - name: Direct Messaging
    description: Endpoints for 1-to-1 direct messaging and conversations
  - name: Group Chat
    description: Endpoints for managing group chats and messages
  - name: User Management
    description: Endpoints for user search and management
paths:
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: User Registration
      description: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successful registration
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: User Login
      description: Authenticate a user and return JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successful login (Returns JWT token)

  /api/dm/conversation:
    get:
      tags:
        - Direct Messaging
      summary: Open Conversation
      description: Open or initialize a 1-to-1 conversation with a peer
      security:
        - bearerAuth: []
      parameters:
        - name: peerUsername
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Conversation opened successfully

  /api/dm/messages:
    get:
      tags:
        - Direct Messaging
      summary: List Messages
      description: Retrieve a list of historical messages with a specific peer
      security:
        - bearerAuth: []
      parameters:
        - name: peerUsername
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: beforeMessageId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of messages retrieved

  /api/dm/recent-conversations:
    get:
      tags:
        - Direct Messaging
      summary: Get Recent Conversations
      description: Retrieve a list of recent direct conversations for the authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Recent conversations retrieved

  /api/groups:
    post:
      tags:
        - Group Chat
      summary: Create Group
      description: Create a new group chat
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                memberUsernames:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Group created successfully

  /api/groups/{groupId}:
    get:
      tags:
        - Group Chat
      summary: Get Group Info
      description: Retrieve information about a specific group chat
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Group info retrieved
    patch:
      tags:
        - Group Chat
      summary: Rename Group
      description: Rename an existing group chat
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Group renamed
    delete:
      tags:
        - Group Chat
      summary: Delete Group
      description: Delete a group chat (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Group deleted

  /api/groups/recent-conversations:
    get:
      tags:
        - Group Chat
      summary: Get Recent Group Conversations
      description: Retrieve a list of recent group conversations for the authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Recent group conversations retrieved

  /api/groups/{groupId}/members:
    post:
      tags:
        - Group Chat
      summary: Add Members
      description: Add new members to an existing group chat
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                usernames:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Members added

  /api/groups/{groupId}/members/{memberUsername}:
    delete:
      tags:
        - Group Chat
      summary: Remove Member
      description: Remove a member from a group chat
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: memberUsername
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Member removed

  /api/groups/{groupId}/leave:
    post:
      tags:
        - Group Chat
      summary: Leave Group
      description: Leave a group chat
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Successfully left the group

  /api/groups/{groupId}/messages:
    get:
      tags:
        - Group Chat
      summary: List Group Messages
      description: Retrieve a list of historical messages in a group chat
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: beforeMessageId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Group messages retrieved

  /api/users/search:
    get:
      tags:
        - User Management
      summary: Search Users
      description: Search for users by username or name
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of matched users

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT